packetbeat/beater: make code more local and add some docs
No logic change.
